refactor(ecs): split runtime.rs into lifecycle/monitoring/lb/config/secrets
Audit 2026-05-19 P8 (partial). 3797-line runtime.rs split: extracted
the 22-method impl EcsRuntime block into 5 concern-based submodules
under runtime/:

- task_lifecycle.rs  641  (run_task/wait_for_task_exit/stop)
- monitoring.rs      190  (health/state changes/log forwarding)
- secrets.rs          47  (resolve_secret)
- config.rs           46  (cli/with_* builders)
- lb.rs               44  (register/deregister LB targets)
- mod.rs            2884  (struct defs, free-standing parse helpers,
                          impl DependsOnCondition, tests)

mod.rs is still the largest file because the bulk of runtime.rs is
free-standing parser functions (build_container_plans, parse_health_check,
parse_ulimit, parse_linux_parameters, ...) plus the test corpus.
Those are concerns-grouped, not impl-method-grouped, so they need a
separate parse_*-keyed split that isn't worth dragging in here.

88 unit tests still pass.
